Script/requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.14.3
boto3==1.42.30
botocore==1.42.30
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cssselect==1.3.0
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.20.3
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-cloud-core==2.5.0
google-cloud-storage==3.8.0
google-crc32c==1.8.0
google-generativeai==0.8.6
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
httplib2==0.31.1
idna==3.11
jieba3k==0.35.1
jmespath==1.0.1
joblib==1.5.3
lxml==6.0.2
lxml_html_clean==0.4.3
newspaper3k==0.2.8
nltk==3.9.2
pillow==12.1.0
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.1.0
regex==2026.1.15
requests==2.32.5
requests-file==3.0.1
rsa==4.9.1
s3transfer==0.16.0
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.8.1
tinysegmenter==0.3
tldextract==5.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
